+++ put_test.go
@@ -33,14 +33,62 @@ func TestDynamoDb_PutItem(t *testing.T) {
 
 	t.Run("PutItemInputRecorded", func(t *testing.T) {
 		tableName := "mockTable"
+		putItemInput := dynamodb.PutItemInput{TableName: &tableName}
+		putItemOutput := dynamodb.PutItemOutput{}
+
+		db := mockdynamodb.NewWithTables([]string{tableName})
+		db.GetTable(tableName).AddReturnPutItemOutput(&putItemOutput)
+
+		output, err := db.PutItem(&putItemInput)
+
+		assert.NoError(t, err)
+		assert.Equal(t, &[]dynamodb.PutItemInput{putItemInput}, db.GetTable(tableName).ReceivedPutItemInputs())
+		assert.Equal(t, &putItemOutput, output)
+	})
+
+	t.Run("PutItemOutputReturnedInOrder", func(t *testing.T) {
+		tableName := "mockTable"
+		putItemInput := dynamodb.PutItemInput{TableName: &tableName}
+
+		putItemOutput1 := dynamodb.PutItemOutput{}
+		putItemOutput2 := dynamodb.PutItemOutput{}
+
+		db := mockdynamodb.NewWithTables([]string{tableName})
+		db.GetTable(tableName).AddReturnPutItemOutput(&putItemOutput1)
+		db.GetTable(tableName).AddReturnPutItemOutput(&putItemOutput2)
+
+		output1, err1 := db.PutItem(&putItemInput)
+		output2, err2 := db.PutItem(&putItemInput)
+
+		assert.NoError(t, err1)
+		assert.NoError(t, err2)
+		assert.Same(t, &putItemOutput1, output1)
+		assert.Same(t, &putItemOutput2, output2)
+	})
+
+	t.Run("PutItemOutputIsNull", func(t *testing.T) {
+		tableName := "mockTable"
+		putItem := dynamodb.PutItemInput{TableName: &tableName}
+
+		db := mockdynamodb.NewWithTables([]string{tableName})
+		db.GetTable(tableName).AddReturnPutItemOutput(nil)
+
+		_, err := db.PutItem(&putItem)
+
+		assert.Error(t, err)
+		assert.Equal(t, &[]dynamodb.PutItemInput{putItem}, db.GetTable(tableName).ReceivedPutItemInputs())
+	})
+
+	t.Run("PutItemOutputEmpty", func(t *testing.T) {
+		tableName := "mockTable"
 		putItem := dynamodb.PutItemInput{TableName: &tableName}
 
 		db := mockdynamodb.NewWithTables([]string{tableName})
 
 		_, err := db.PutItem(&putItem)
 
-		require.NoError(t, err)
-		assert.Equal(t, &[]dynamodb.PutItemInput{putItem}, db.GetTable(tableName).Items())
+		assert.Error(t, err)
+		assert.Equal(t, &[]dynamodb.PutItemInput{putItem}, db.GetTable(tableName).ReceivedPutItemInputs())
 	})
 }