+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.1] - 2021-02-14
+### Added
+- Query() implemented
+
+### Changed
+- How inputs and outputs for PutItem() work
+
 ## [0.0.0] - 2021-02-04
 ### Added
 - PutItem() implemented